Removed custom wrapper files, added Janus DiffusionPipeline wrapper to model_utils, cleaned up configs
